test(markerscope): add comprehensive tests for FieldOrType markers
